Allow using Maps to specify nested annotations
All Java Annotations extend java.lang.annotation.Annotation,
which defines a method `annotationType` that return the type
of the annotation (and not the proxy class). As such, it can
be used as a key in a Map to specify the type of the Annotation
to create, with the remaining keys being the elements of the
Annotation.