Add nested AnnotationCreator alternative; add Javadoc for AnnotationCreator
Add a new method to support creating nested annotations using
AnnotationCreators. It works by creating a new AnnotationCreator
and adding its value map as a value to the parent AnnotationCreator
(thus any changes made to the nested AnnotationCreator update the
map in the parent's values). Also added Javadoc so these features
are discoverable.