Small change to how `float` and `double` fields are compared in generated `equals()`
The `Gizmo.generateEquals()` method used to emit direct `==` comparison
even for `float` and `double` fields. This is inconsistent with how
arrays of these types are compared (`Arrays.equals()`) and also with
the template code IntelliJ generates.

This commit changes that to bitwise comparison. Specifically, `float`
comparison is now an `int` comparison of `Float.floatToIntBits`, and
`double` comparison is `long` comparison of `Double.doubleToLongBits`.