Introduce ClassTransformer
The `ClassTransformer` can:

- add implemented interfaces
- add and remove modifiers
- add fields and methods
- modify fields and methods

To modify fields and methods, extra classes called `FieldTransformer`
and `MethodTransformer` exist. They can:

- add and remove modifiers
- rename

More features may be added relatively easily.

A `ClassTransformer` must be applied to a `ClassVisitor` that has not been
visited yet. If applied to an already used visitor, an exception is thrown
in `visitEnd()`.

Co-authored-by: Ladislav Thon <lthon@redhat.com>