fix primitive constants
We need constants of types `byte`, `short` and `char` because without them,
we cannot statically make a difference between them and `int`, which may
lead to all kinds of issues. For example, we could emit an invocation of
a different overload of a method than expected, if the target class has
an overload of the method for both `byte` (or `short` or `char`) and `int`.

In addition to that, this commit also adds `Constant.of(Boolean)`, because
`java.lang.Boolean` constants were not of the primitive type, unlike all
others. This might be a bad idea, but all other constants of wrapper types
are in fact constants of the corresponding primitive type, and consistency
is warranted. We might later want to rethink that.