fix the ArrayStore item
The `ArrayStore` item used to emit an array store instruction based on
the type kind of the array, which is always reference. It needs to do
that based on the type kind of the _component type_ of the array.

Further, this commit replaces all occurences of "element type" with
"component type", because that's what they really are. This doesn't
make a difference with single-dimensional arrays, which is the vast
majority of arrays, but it does make a difference with multi-dimensional
arrays, where component type is the type of the inner array, while
the element type is never an array type.

Finally, some tests for arrays are added.