track added members in TypeCreator
For now, we only remember descriptors of added members (instance/static
fields, instance/static methods, constructors) and provide public access
to them.

In the future, we might want to remember all added creators, to be able
to prevent adding the same member multiple times, but even in that case,
we should still only expose access to descriptors.