improve the Expr type hierarchy
The previous `Expr` type hierarchy looked like this:

Expr
├─ Constant
├─ LValueExpr
│  ├─ FieldVar
│  │  ├─ InstanceFieldVar
│  │  └─ StaticFieldVar
│  ├─ LocalVar
│  ├─ ParamVar
│  └─ LValueExprImpl
├─ Var
│  ├─ FieldVar
│  │  ├─ StaticFieldVar
│  │  └─ InstanceFieldVar
│  ├─ LocalVar
│  ├─ ParamVar
│  └─ ThisExpr
└─ Item
   ├─ ThisExpr
   └─ other subclasses

The new `Expr` hierarchy looks like this:

Expr
├─ Constant
├─ This
│  └─ ThisExpr
├─ Assignable (previously `LValueExpr`)
│  ├─ Var
│  │  ├─ FieldVar
│  │  │  ├─ InstanceFieldVar
│  │  │  └─ StaticFieldVar
│  │  ├─ LocalVar
│  │  └─ ParamVar
│  └─ AssignableImpl (previously `LValueExprImpl`)
└─ Item (many subclasses also implement the interfaces above)

All `Var`s are `Assignable`s now. `ThisExpr` is no longer a `Var`,
it has a separate user-facing interface `This`. It no longer has
a `name()` method, which was the only method prescribed by the `Var`
interface. I don't think this is a big loss.