Do not intern descriptors and constants
These objects are usually either short lived (e.g. the lifespan of method creation) or very long lived (e.g. stored in a constant field). Limiting them to the lifespan of the Gizmo instance does not make a lot of sense in this perspective. But, leave the door open to allow us to later explicitly intern these constants.