refactor flags handling on ExecutableCreators
The `ExecutableCreatorImpl` constructor takes two sets of flags: default
flags, which is the starting point, and allowed flags, which are the flags
that may be added. The default flags may not be removed.

Then, the `withFlag()` method checks that the added flag is present
in the set of allowed flags, while the `withoutFlag()` method checks that
the removed flag is not present in the set of unremovable flags.

No subclass overrides the `withFlag()` method anymore, they just pass
the correct sets to the superclass constructor.

All methods related to adding/removing flags are now `final`.