refactor flags handling on FieldCreators
The `FieldCreatorImpl` constructor takes the set of default flags,
which is the starting point. The default flags may not be removed.

Then, the `withoutFlag()` method checks that the removed flag is
not present in the set of unremovable flags.

In contrast to `ExecutableCreatorImpl`, there's no explicit set of
allowed flags. The existing check in `withFlag()` is good enough.

All methods related to adding/removing flags are now `final`.