add tests for arithmetic operations and fix type check
The type check in `BinOp` used to assume that all binary operations
accept operands of the same type. That is not true: the shifts
(shl, shr, ushr) always accept an `int` second operand, even if
the first operand is `long`. This issue is fixed now.

Further, this commit adds tests for arithmetic operations on `int`,
`long` and `double`.