check arguments count and types when invoking methods
We first check that the number of passed arguments is equal to the number
of declared parameters. Then, we check that `TypeKind.asLoadable()` of
all arguments are equal to the loadable `TypeKind` of the corresponding
parameters. This is technically relatively lax and only properly checks
primitive types, but we don't have a type system to find out if references
of different types (such as `String` and `CharSequence`) are assignable.
Therefore, this is the best we can do (at least for now).