the API to create annotations no longer leaks the ClassFile API
The methods to add nested annotation[s] no longer attempt to load the class
of the nested annotation from a classloader. Instead, they take it from
the return type of the annotation element.

Additionally, the annotation creation API now allows creating annotations
based on a `ClassDesc` description of the annotation type, not only based
on the `Class<? extends Annotation>` description.