Change organization of `Gizmo` and `ClassOutput`
This more closely aligns with Gizmo 1, and simplifies the API for class creation, but still leaves room to implement #222 in some sane way.

It stops short of using static methods, but with static methods it becomes difficult to establish a configuration that would persist but have constrained scope, like what would be required for #222.