add ConstructorDesc.of() overloads that accept vararg parameter types
It is often needed to invoke a constructor of a dynamically created class
which has a statically known parameter list. The new overloads are easier
to use in such situation.

It is unfortunately not possible to add an overload that would accept
a `List<Class<?>>`, because that would have the same erasure as an already
existing method that takes `List<ClassDesc>`.