optionally track Item creation sites and improve the exception message
When the system property `gizmo.trackCreations` is set (even to an empty
value, the only exception is the `false` value), all created `Item`s
will carry a textual description of their creation site (which consists
of the class name, method name and line number). When the Item is used
on an unexpected place, that creation site description is used as part
of the exception message, greatly improving the debugging experience.