rename Constant to Const
We don't shy away from using abbreviations for very common concepts, such as
`Expr` (expression), `Var` (variable) or `Desc` (descriptor). Constants are
very common as well, and "const" is a common abbreviation of "constant" (so
common that it is in fact a Java keyword!), so that's what we should use.

All the other related classes (such as `ConstantImpl` and all its subclasses)
are renamed too.