rename BlockCreator.redo() to various form of jumpTo*()
Specifically, this is:

- `redo(BlockCreator)` -> `jumpToBlock()`
- `redo(SwitchCreator, Const|Class|Enum|String|int)` -> `jumpToCase()`
- `redo(SwitchCreator)` -> `jumpToDefault()`
- `redo()` -> `jumpToStart()`

The term "redo" is not common in the Java lingo. The new names are
more descriptive and use common wording. We might want to replace
the verb "jump" with "goto", which is an actual Java keyword.

The implementation classes are renamed as well. As a special case,
the base class is renamed from `Goto` to `Jump`. If we decide to use
the verb "goto" instead of "jump", this would be renamed back.