rename BlockCreator.selectExpr() to cond()
The term "select" is not common in the Java world. The closest Java
cousin, Kotlin, uses the term to wait for multiple coroutines at once
and pick the result of the first completion. In Go, the `select` keyword
has practically the same meaning: wait for multiple channels at once
and select the first completion. (In Erlang, for constrast, this
operation is called `receive`.)

The meaning of `selectExpr()` was to evaluate the conditional operator,
so the best corresponding name is `conditional()` or `cond()` for short.
I wouldn't be against using the full name, but went with the short one.