`unsafeCast` -> `uncheckedCast`
On further thought I've decided that `uncheckedCast` is a better name.
Everyone knows what "unchecked" means - it means the cast isn't checked.
The cast is not "unsafe" in any way - if the cast is valid, it's safe, and if it's invalid, it will fail verification.
It also intuitively complements the `CHECKCAST` instruction.