replace List<Expr> by List<? extends Expr> in BlockCreator methods
We have a few public types that extend `Expr` and users can easily
have `List`s of them. For example, if `method` is a Jandex
`MethodInfo`, here's how to create a list of parameters of a method
that is supposed to forward the call to the `method`:

```java
List<ParamVar> params = IntStream.range(0, method.parametersCount())
        .mapToObj(i -> mc.parameter("param" + i, i))
        .toList();
```

However, the `invoke*()` methods on `BlockCreator` didn't accept
`List<ParamVar>`, even though it is perfectly valid. This commit
fixes that, and a lot of other similar cases.