improve caller tracking message so that IDEs show a code navigation link
The output format is specifically tested with IntelliJ IDEA, which converts
the <file name>:<line number> part of the message to a code navigation link.
Not sure about other IDEs, but the format is the same as in exception stack
traces, so it's likely other IDEs recognize it as well.