fix Item.pop() for New and NewEmptyArray
The `New` and `NewEmptyArray` items are at the beginning of the list
of instructions that produce the ultimate result, so `Item.pop()` doesn't
work correctly. That is because `Item.pop()` inserts the `Pop` instruction
right after the item in the list, which means that a bunch of other items
that follow `New` / `NewEmptyArray` remain intact.

In this commit, we fix that by creating new items that represent the result
of `New` / `NewEmptyArray` (called `NewResult` / `NewArrayResult`). These
items are at the end of the list, so inserting `Pop` after them works
correctly.