Rework item binding
An item is bound if it is added to the list. Constants are never bound. Derefs are bound once they have been added iff the thing they refer to is bound. Bound items may be used only once. Derefs of constants (or derefs of deref of constants etc.) can be reused many times.

Fixes #372