Create flags and access levels abstraction
Fixes #335.

Version 1 of this change used type-safe constants for all of the modifiers and access levels; it was pretty nifty but added about 30-40 new types, so it seemed a bit to heavy to justify.

This version makes a fair effort to put convenience methods in the right place, but otherwise does all checking at run time.

At present, if you try to introduce a flag to a creator which always forces that flag to be set, then you will get an exception. I'm undecided whether this is a preferable behavior.

I did not do anything about default access levels or modifiers yet, other than more or less mirroring what was already in the code.