remove GenericType and related classes
This commit moves back to a single, non-generic representation of types
using `ClassDesc`. This is faithful to bytecode, which only retains
generics in the `Signature` attribute, but is otherwise non-generic.

The introduction of `GenericType` increased complexity (two representations
of types instead of one) and irregularity (you sometimes have to use
`GenericType`, and other times, you have to use `ClassDesc`), while only
providing mediocre benefits (signatures are generated automatically and
type annotations are possible; on the other hand, you can create a field
of type `List<String>`, but when you read from it, you get a `List`).
To really gain the benefits, we would have to represent types using
`GenericType` _everywhere_, including field/method/type descriptors.
That would be a significant undertaking and the API would be harder
to use (especially with current `GenericType`, which is relatively hard
to construct).