add specialized API to set class/method/field signatures
The API includes:

- representation of generic types (`GenericType` and all its member types)
- signature creators (`ClassSignatureCreator`, `InterfaceSignatureCreator`,
  `MethodSignatureCreator`, `FieldSignatureCreator`)

Using the API leads to certain amount of information duplicity. This is
unfortunate, but it's the price for not spreading the generic type API
across all of the code base.

This API doesn't include type annotations, which should be relatively
straightforward to add if there's a desire.