improve the TypeArgument type hierarchy
The previous type hierarchy was:

```
TypeArgument
├─ OfAnnotated
│  ├─ OfExtends
│  ├─ OfSuper
│  └─ Wildcard
└─ OfExact
```

and included one intermediary interface:

```
OfBounded (*)
├─ OfExact
├─ OfExtends
└─ OfSuper

(*) the interface
```

This is modeled after the `Signature` type hierarchy in the ClassFile API,
which is in several aspects different to how types are treated in the JLS
and how people commonly think about them.

The new type hierarchy attempts to solve this by modelling type arguments
closely to how the JLS treats them:

```
TypeArgument
├─ OfExact
└─ OfWildcard
   ├─ OfExtends
   ├─ OfSuper
   └─ OfUnbounded
```

And includes the following hierarchy of intermediary interfaces:

```
OfTyped (*)
├─ OfExact
└─ OfBounded (*)
   ├─ OfExtends
   └─ OfSuper

(*) the interfaces
```

Further, this commit adds missing `equals()` and `hashCode()` methods
to `TypeArgument` and all its subclasses.

Finally, with this commit, the unbounded wildcard is presented as `?`
in `toString()`, instead of `*` (which is how it appears in signatures).