fix TypeVariable.erasure()
This method assumed that if there's no first bound, then the erasure
of the type variable is always `java.lang.Object`, but that doesn't
necessarily have to be true. It is possible for a type variable to have
no first bound, but have some other bounds -- in case all bounds are
interface types. The erasure of a type variable is the erasure of its
leftmost bound, which in this case is the first of the other bounds.