rename TypeVariable to TypeParameter
This is because there's a difference between type variables and type parameters:
type parameters may declare bounds. We already had that difference in our
types: `TypeVariable` vs. `GenericType.OfTypeVariable`. This rename makes that
difference more apparent and the names less confusing.

Also, `TypeVariableCreator` is renamed to `TypeParameterCreator`, and a few more
places related to that.

Finally, the `TypeParameterizedCreator.typeParameter()` method no longer returns
a `TypeParameter`; it returns `GenericType.OfTypeVariable` now. This is what can
be used directly when creating a `GenericType`.