fix how lambdas are generated
To generate a lambda, we create an anonymous class, serialize it
and use Base64 to encode the serialized class. (The invokedynamic
bootstrap method then decodes the serialized form and spins up
a new class.) We used the default Base64 alphabet, which includes
a disallowed character: `/`.

With this commit, we use the URL-safe Base64 alphabet, which does
not include any disallowed characters.

This commit also adjusts the error message for not allocated local
variables: this may also happen if a lambda uses a non-captured
variable from outer lexical scope. Took me a while to figure out,
so I figured the error message should contain a hint.