add ExecutableCreator.parameter() methods that accept GenericType
Previously, it was only possible to declare a parameter of a `GenericType`
by using the `ParamCreator`. This is unwieldy, especially when creating
a forwarding method that needs to declare parameters of the same type
as an original method. To do that, we need to access the original parameter
type in a loop using the loop index, but that isn't `final` and so cannot
be used directly in the `Consumer<ParamCreator>` lambda. The new methods
added in this commit make this straightforward.