only put the synthetic flag on generated classes/interfaces
This is to workaround the IntelliJ IDEA decompiler (FernFlower) deficiency
where it ignores all synthetic members of decompiled classes. Since we used
to mark all members as synthetic, a decompiled class would be just empty,
which is certainly not good.

We still mark the generated classes/interfaces as synthetic, which should
be good enough of a marker to everyone that all members are also synthetic.