centralize the counter for generating anonymous class and lambda names
Previously, the counter was present in `BlockCreatorImpl` and so was unique
for each block. This is not necessarily correct; it is possible for one method
to have multiple blocks and each of those blocks to generate an anonymous
class. With this commit, the counter is present in `TypeCreatorImpl` and so
is unique to each generated type. This is always correct.