centralize the counter for generating anonymous classes and lambdas
Previously, the counter was present in `BlockCreatorImpl` and so was unique
for each block. This is not necessarily correct; it is possible for one method
to have multiple blocks and each of those blocks to generate an anonymous
class. With this commit, the counter is present in `TypeCreatorImpl` and so
is unique to each generated type. This is always correct.

Also, the translation strategy for lambdas used to generate classes with a name
derived only from the name of the enclosing class. This does not necessarily
have to be wrong, but this commit adds a counter value to that name too, which
is arguably more correct.