fix numeric promotion for eq/ne comparisons
The non-equality comparisons (`<`, `<=`, `>`, `>=`) require numeric promotion
always, but equality comparisons (`==`, `!=`) only require it if at least one
argument is numeric type (that is, primitive but not `boolean`). If both
arguments are not numeric types, no numeric promotion should happen.

We previously guarded against that in a naive way: the `numericPromotion()`
method returned early when both operands were of the same type after unboxing.
This leads to incorrect conversion in case of equality comparisons on primitive
wrapper types (and may lead to NPE at runtime). This commit adds proper guard
to the appropriate place.

Note that there are other occurences of `numericPromotion()` which are not
guarded by `numericPromotionRequired()`. These are in `BinOp` and in `Cmp`,
where on both places the operation is always numeric and so always requires
numeric promotion.