improve lambda method names in debug mode
Previously, lambda methods in debug mode were always named `$$lambda$N`,
where `N` is a counter. Now, they are named similarly to how javac names
them: `lambda$<name>$N`, where `<name>` is the declaring method name
(with `static` for static initializers and `new` for constructors) and
`N` is a counter. This makes it easier to figure out where exactly
the lambda is created.