add reference narrowing conversion, but only from java.lang.Object
Such conversion is not type safe, but it is incredibly useful. I tried to
make it opt-in (by adding `Gizmo.withObjectDowncasts()`) and it mostly worked,
except I couldn't pass the `Gizmo` object to `ArrayDeref` and `ArrayStore`.
Those would require all `Item`s to have an instance of `Gizmo`, which is not
possible for `Const`s. Hence, it is enabled always, which is not terribly
bad, because Gizmo 1 allowed it too.