make Gizmo 2 a multimodule project
With the anticipation of adding a 2nd module, Jandex integration, this
commit turns Gizmo 2 into a multimodule Maven project. The main code
is now part of the `core` module, which retains the previous Maven
coordinates as well as the JPMS module name.

The top-level project is a Maven `pom` module which contains no code.