add Jandex integration
The `jandex` module is added which hosts the Jandex integration. It depends
on the `core` module (both in the Maven sense and in the JPMS sense) and
of course on Jandex.

The `Jandex2Gizmo` class contains `static` methods to create `ClassDesc`
(and other descriptors) from Jandex types (such as `DotName`, `Type` or
`ClassInfo`).

More methods will be coming in the future (notably, to create `GenericType`
from `Type` and to create annotation from `AnnotationInstance`).