fix try/finally in case the try body has nested blocks
The nested blocks in the `try` body block did _not_ have the `tryFinally`
field set, because that field is only set when the `finally` block is
added. When the `try` body block is being generated, the field is `null`
and hence copying that in the block constructor is _not_ enough to correctly
propagate the enclosing `TryFinally`.

With this commit, we no longer attempt to propagate the enclosing `TryFinally`
in the block constructor; we only let it be set on the outermost try body block.
Instead, inside `writeCode()`, which is the only place we need it, we look it up
in the block hierarchy.