fix StringBuilderOps
The `StringBuilderOps` class didn't guarantee that the object it operates
upon is stored in a local variable. This was previously partially fixed by
all the `public` methods returning a new instance, but that requires usages
to remember that and always reassign. This is error-prone and inconvenient.

With this commit, all constructors of `StringBuilderOps` now guarantee that
the object is indeed stored in a local variable, which makes `return this`
a valid usage pattern in all the `public` methods. Usage is then easy and
less error-prone.