add ClassHierarchyLocator and use it to create ClassFile with ClassHierarchyResolver
To be able to generate stack maps, the `ClassFile` API uses
a `ClassHierarchyResolver`. By default, it only has access to classes
loaded by the system classloader, which is not good enough.

In this commit, we add a `ClassHierarchyLocator` that can be set on `Gizmo`.
A `ClassHierarchyLocator` is our variant of `ClassHierarchyResolver` which
includes several factory methods and helpers to easily create a result.

When the `ClassHierarchyLocator` is set on `Gizmo`, we use it to construct
`ClassFile` with a `ClassHierarchyResolver` that has access to all classes
known to the given `ClassHierarchyLocator`. When it is not set, we don't add
a `ClassHierarchyResolver` to the `ClassFile`.