create ClassFile with ClassHierarchyResolver
To be able to generate stack maps, the `ClassFile` API uses
a `ClassHierarchyResolver`. By default, it only has access to classes
loaded by the system classloader, which is not good enough.

In this commit, we provide a `ClassHierarchyResolver` that has access
to all classes loaded by the TCCL as set when `ClassFile` is being
created. Only if there is no TCCL we fallback to the system CL.