Avoid executing TypeKind.from() too much by caching the value
In this commit, I'm doing it surgically in places where it's important
for ArC but tbh I wonder if it should be handled at a more generic level
i.e. maybe we should have a Item(ClassDesc type) constructor and
enforce all subclasses to define their type this way (and cache the
TypeKind in Item).