make withStringBuilder() accept Var and document reusability of Ops instances
Instances of nearly all `*Ops` classes may _not_ be reused: only a single
method may be called on the result of `BlockCreator.with*()`.

The exception is `StringBuilderOps`, which is explicitly designed to allow
reusing in order to mimic usage of `StringBuilder`. To highlight reusability,
the `withStringBuilder()` method now accepts a `Var` instead of `Expr`.
Local variables are not allocated behind the scenes; they must be provided
from the outside.

The `StringBuilderOps` class now has a single constructor that accepts
a local variable as well.

The `withNewStringBuider()` methods are not changed, they declare a local
variable internally.