more complete verification of calls on enclosing BlockCreators
We already had a facility to verify that a `BlockCreator` whose "child"
is currently being created cannot be used. This facility was very limited
though -- it only applied in the `block()` and `blockExpr()` methods.
This commit expands that to all methods that contribute to the same
method, such as conditions, loops or exception handlers. A test is added
that verifies all these.

Note that we do _not_ attempt to verify this across multiple methods,
including lambdas and anonymous classes. This would be more costly.