fix emitting nests
The anonymous classes we generate have the `NestHost` attribute, but
the owning class does _not_ have the `NestMembers` attribute. Not sure
if that is actually checked, but per the JVMS, both attributes must
be present. This commit adds the `NestMembers` attribute to the owning
class, listing all generated anonymous classes.