optimize `BlockCreator.exprEquals()` / `exprHashCode()` / `exprToString()`
In case the argument (or at least one of the arguments in case of `exprEquals()`)
is a non-null reference constant, we don't have to call the null-safe methods
we usually call: `Objects.equals()`, `Objects.hashCode()`, `String.valueOf()`.
Instead, we can directly call `Object.equals()` / `hashCode()` / `toString()`.
This commit does exactly that.